Update Spot API endroute

+++ backend/db/migrations/20220927013524-create-review.js
@@ -29,7 +29,7 @@ module.exports = {
         allowNull:false
       },
       stars: {
-        type: Sequelize.DECIMAL(2,1),
+        type: Sequelize.INTEGER,
         allowNull:false
       },
       createdAt: {
